Revert "Fix bug in FordFulkerson.hpp"
This reverts commit b83ae4c4c60428da621e553e54c51186cb115b51.

+++ include/FordFulkerson.hpp
@@ -52,7 +52,7 @@ void FordFulkerson(GT& G, umap_WT& c, T s, T t, umap_WT& f) {
 		}
 		umap<T, BFSInfo<T>> BFS_ans;
 		BFS(Gf, s, BFS_ans);
-		std::vector<int> p;
+		std::vector<size_t> p;
 		PrintPath(s, t, BFS_ans, p);
 		if (!p.size())
 			break;
